requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
django-extensions==3.2.3
psycopg2-binary==2.9.6
python-decouple==3.8
Pillow==9.5.0
djangorestframework-simplejwt==5.2.2
celery==5.2.7
redis==4.5.5
gunicorn==20.1.0
whitenoise==6.4.0
